backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
